docs: create Linux ABI compatibility guide and update documentation with implementation findings (TEAM_217)
- Add `docs/specs/LINUX_ABI_GUIDE.md` documenting critical ABI alignment knowledge (Stat layout, auxv, TLS context switching)
- Update `docs/ARCHITECTURE.md` to link new guide and userspace ABI spec
- Update `docs/specs/userspace-abi.md` to v1.1.0 with corrected Stat structure, syscall status, and auxv specification
- Implement TLS (`TPIDR_EL0`) context switching in `cpu_switch_to` assembly

+++ src/syscall/time.rs
@@ -83,8 +83,8 @@ pub fn sys_clock_gettime(timespec_buf: usize) -> i64 {
         let remainder = counter % freq;
         let nanoseconds = (remainder * 1_000_000_000) / freq;
         Timespec {
-            tv_sec: seconds,
-            tv_nsec: nanoseconds,
+            tv_sec: seconds as i64,
+            tv_nsec: nanoseconds as i64,
         }
     } else {
         Timespec::default()
